% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_alg.R
\name{detec_otus}
\alias{detec_otus}
\title{An algrithm for phylogeny-aware differential abundance detection}
\usage{
detec_otus(p.val, tree_table, tree, group, test.method, cutf, adj.m)
}
\arguments{
\item{p.val}{the p value for each node}

\item{tree_table}{a data set}

\item{tree}{a phylogenetic tree among m taxa}

\item{group}{a n-vector of group indicators}

\item{test.method}{t-test or Wilcoxon rank sum test}

\item{cutf}{level of significance}

\item{adj.m}{the adjustment methods for p-values}
}
\value{
final.p the adjusted p values

dif.otus  the set of differentially abundant OTUs
}
\description{
An algrithm for phylogeny-aware differential abundance detection
}
\examples{
p <- 40
set.seed(1)
tree <- simulate_tree(p)
leaf_ind <- leaf_node_index(tree)
}
