% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.R
\name{eBay}
\alias{eBay}
\title{An method for differential abundance detection}
\usage{
eBay(otu.data, group, test.method, cutf)
}
\arguments{
\item{otu.data}{the data set where each row was a sample and each column was a taxa,}

\item{test.method}{a test including "t" and "wilcoxon"}

\item{cutf}{the significance level}
}
\value{
dif.otus the detected differential otus

final.p the p value for each otu
}
\description{
An method for differential abundance detection
}
\examples{
group <- rep(c(0,1),each=46)
ebay.res <- eBay(otu.table,group,test.method="t",cutf=0.05)
}
