% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.tree.R
\name{eBay_tree}
\alias{eBay_tree}
\title{An method for phylogeny-aware differential abundance detection
the eBay incorporates the tree into the normalization process and implement
differential abuance testing.}
\usage{
eBay_tree(otu.data, tree, group, test.method, cutf, adj.m)
}
\arguments{
\item{otu.data}{the data set where each row was a sample and each column was a taxa,}

\item{tree}{a phylogenetic tree for otus}

\item{test.method}{a test including "t" and "wilcoxon"}

\item{cutf}{the significance level}
}
\value{
dif.node the detected differential nodes

p.value the p value for each taxa
}
\description{
An method for phylogeny-aware differential abundance detection
the eBay incorporates the tree into the normalization process and implement
differential abuance testing.
}
