% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_alg.R
\name{leaf_node_index}
\alias{leaf_node_index}
\title{An algrithm for phylogeny-aware differential abundance detection}
\usage{
leaf_node_index(tree)
}
\arguments{
\item{tree}{a phylogenetic tree for otus}

\item{p.val}{the p value for each node}

\item{test.method}{a test including "t" and "wilcoxon"}

\item{tree_table}{a data set}

\item{cutf}{the significance level}

\item{adj.m}{the adjustment methods for p-values}
}
\value{
dif.node the detected differential nodes

p.value the p value for each taxa
}
\description{
An algrithm for phylogeny-aware differential abundance detection
}
\examples{
p <- 40
set.seed(1)
tree <- simulate_tree(p)
leaf_ind <- leaf_node_index(tree)
}
