% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.tree.R
\name{otu_index}
\alias{otu_index}
\title{An method for phylogeny-aware differential abundance detection
the eBay incorporates the tree into the normalization process and implement
differential abuance testing.}
\usage{
otu_index(tree)
}
\arguments{
\item{tree}{a phylogenetic tree for otus}

\item{otu.data}{the data set where each row was a sample and each column was a taxa,}

\item{group}{the group label for each sample}

\item{g}{the internal node of the tree}

\item{tree_table}{the total data set including the internal nodes and leaf nodes count}

\item{test.method}{a test including "t" and "wilcoxon"}

\item{cutf}{the significance level}

\item{adj.m}{the adjustment methods for p-values}
}
\value{
dif.node the detected differential nodes

p.value the p value for each taxa
}
\description{
An method for phylogeny-aware differential abundance detection
the eBay incorporates the tree into the normalization process and implement
differential abuance testing.
}
\examples{
####generate data####
p <- 40
set.seed(1)
tree <- simulate_tree(p)
set.seed(1)
control_pi = case_pi = c()
for(j in (p+1):(p+tree$Nnode)){
set.seed(j)
random_pi <- runif(1,0.2,0.4)
control_pi[which(tree$edge[,1]==j)] <- c(random_pi, 1-random_pi)
case_pi[which(tree$edge[,1]==j)] <- c(random_pi, 1-random_pi)}
control_theta = case_theta = rep(0.1, tree$Nnode)
group <- rep(c(0,1),each =20)
tree_table <- simulation_dtm(p=40,tree, seed=1, N=20,control_pi, case_pi,
control_theta,case_theta)
#####differential abundance testing based on a phylogenetic tree###
ebay_tree.res <- eBay_tree(otu.data=tree_table, tree=tree, group=group,
test.method="t", cutf=0.05,adj.m="BH")
}
