% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tree.R
\name{simulate_tree}
\alias{simulate_tree}
\title{simulate data from DTM}
\usage{
simulate_tree(n)
}
\arguments{
\item{n}{the number of taxa}
}
\description{
simulate data from DTM
}
\examples{
p <- 40
set.seed(1)
tree <- simulate_tree(p)
}
