% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dm.R
\name{simulation_dm}
\alias{simulation_dm}
\title{simulate data from DM}
\usage{
simulation_dm(p, seed, N, control_pi, case_pi, control_theta, case_theta)
}
\arguments{
\item{p}{the number of taxa}

\item{seed}{a random seed}

\item{N}{sample size of each group}

\item{control_pi}{a probability vector for each taxon which were sum to 1 in control group}

\item{case_pi}{a probability vector for each taxon which were sum to 1 in case group}

\item{control_theta}{the over-dispersion parameter}

\item{case_theta}{the over-dispersion parameter}
}
\value{
ntree_table the otu table
}
\description{
simulate data from DM
}
\examples{
set.seed(1)
rand_pi <- runif(20)
control_pi = case_pi = rand_pi/sum(rand_pi)
control_theta = case_theta = 0.1
group <- rep(c(0,1),each =20)
ntree_table <- simulation_dm(p=20,seed=1, N=20,control_pi, case_pi,control_theta,case_theta)
}
