% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dtm.R
\name{simulation_dtm}
\alias{simulation_dtm}
\title{simulate data from DTM}
\usage{
simulation_dtm(
  p,
  seed = 1,
  N,
  tree,
  control_pi,
  case_pi,
  control_theta,
  case_theta
)
}
\arguments{
\item{p}{the number of taxa}

\item{seed}{a random seed}

\item{N}{sample size of each group}

\item{tree}{a phylogenetic tree}

\item{control_pi}{a probability vector for each taxon which were sum to 1 in control group}

\item{case_pi}{a probability vector for each taxon which were sum to 1 in case group}

\item{control_theta}{the over-dispersion parameter}

\item{case_theta}{the over-dispersion parameter}
}
\value{
dtm_table the otu table
}
\description{
simulate data from DTM
}
\examples{
p <- 40
set.seed(1)
tree <- simulate_tree(p)
control_pi = case_pi = c()
for(j in (p+1):(p+tree$Nnode)){
set.seed(j)
random_pi <- runif(1,0.2,0.4)
control_pi[which(tree$edge[,1]==j)] <- c(random_pi, 1-random_pi)
case_pi[which(tree$edge[,1]==j)] <- c(random_pi, 1-random_pi)}
control_theta = case_theta = rep(0.1, tree$Nnode)
group <- rep(c(0,1),each =20)
tree_table <- simulation_dtm(p=40,tree, seed=1, N=20,control_pi,
case_pi,control_theta,case_theta)
}
