% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.R
\name{eBay}
\alias{eBay}
\title{An method for differential abundance detection}
\usage{
eBay(otu.data, group, cutf, test.methods, adj.m)
}
\arguments{
\item{otu.data}{an OTU table with n rows (samples) and m columns (taxa)}

\item{group}{a n-vector of group indicators}

\item{cutf}{level of significance}

\item{adj.m}{the adjustment methods for p-values}

\item{test.method}{t-test or Wilcoxon rank sum test}
}
\value{
final.p    the adjusted p values

dif.otus   the set of differentially abundant OTUs
}
\description{
An method for differential abundance detection
}
\examples{
####generate data####
library(eBay)
set.seed(1)
rand_pi <- runif(20)
control_pi = case_pi = rand_pi/sum(rand_pi)
control_theta = case_theta = 0.1
group <- rep(c(0,1),each =20)
ntree_table <- simulation_dm(p=20,seed=1, N=20,control_pi, case_pi,
control_theta,case_theta)
#####differential abundance testing###
ebay.res <- eBay(otu.data=ntree_table, group=group, test.method="t",
cutf=0.05,adj.m="BH")
}
