% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.R
\name{eBay}
\alias{eBay}
\title{An method for differential abundance detection}
\usage{
eBay(otu.data, group, cutf, test.methods, adj.m)
}
\arguments{
\item{otu.data}{the data set where each row was a sample and each column was a taxa,}

\item{cutf}{the significance level}

\item{test.method}{a test including "t" and "wilcoxon"}
}
\value{
dif.otus   the detected differential otus

final.p    the p value for each otu
}
\description{
An method for differential abundance detection
}
