% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.R
\name{eBay}
\alias{eBay}
\title{An method for differential abundance detection}
\usage{
eBay(otu.data, group, cutf, test.methods, adj.m)
}
\arguments{
\item{otu.data}{the data set where each row was a sample and each column was a taxa,}

\item{cutf}{the significance level}

\item{test.method}{a test including "t" and "wilcoxon"}
}
\value{
dif.otus   the detected differential otus

final.p    the p value for each otu
}
\description{
An method for differential abundance detection
}
\examples{
library(eBay)
set.seed(1)
rand_pi <- runif(20)
control_pi = case_pi = rand_pi/sum(rand_pi)
control_theta = case_theta = 0.1
group <- rep(c(0,1),each =20)
ntree_table <- simulation_dm(p=20,seed=1, N=20,control_pi, case_pi,
control_theta,case_theta)
ebay.res <- eBay(otu.data=ntree_table, group=group, test.method="t", cutf=0.05)
}
