% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.tree.R
\name{eBay_tree}
\alias{eBay_tree}
\title{An method for phylogeny-aware differential abundance detection
the eBay incorporates the tree into the normalization process and implement
differential abuance testing.}
\usage{
eBay_tree(otu.data, tree, group, test.method, cutf, adj.m)
}
\arguments{
\item{otu.data}{an OTU table with n rows (samples) and m columns (taxa)}

\item{tree}{a phylogenetic tree among m taxa}

\item{group}{a n-vector of group indicators}

\item{test.method}{t-test or Wilcoxon rank sum test}

\item{cutf}{level of significance}

\item{adj.m}{the adjustment methods for p-values}
}
\value{
final.p the adjusted p values

dif.otus  the set of differentially abundant OTUs
}
\description{
An method for phylogeny-aware differential abundance detection
the eBay incorporates the tree into the normalization process and implement
differential abuance testing.
}
\examples{
####generate data####
p <- 40
set.seed(1)
tree <- simulate_tree(p)
set.seed(1)
control_pi = case_pi = c()
for(j in (p+1):(p+tree$Nnode)){
set.seed(j)
random_pi <- runif(1,0.2,0.4)
control_pi[which(tree$edge[,1]==j)] <- c(random_pi, 1-random_pi)
case_pi[which(tree$edge[,1]==j)] <- c(random_pi, 1-random_pi)}
control_theta = case_theta = rep(0.1, tree$Nnode)
group <- rep(c(0,1),each =20)
tree_table <- simulation_dtm(p=40,tree, seed=1, N=20,control_pi, case_pi,
control_theta,case_theta)
#####differential abundance testing based on a phylogenetic tree###
ebay_tree.res <- eBay_tree(otu.data=tree_table, tree=tree, group=group,
test.method="t", cutf=0.05,adj.m="BH")
}
