% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay.tree.R
\name{otu_index}
\alias{otu_index}
\title{An method for phylogeny-aware differential abundance detection}
\usage{
otu_index(tree)
}
\arguments{
\item{tree}{a phylogenetic tree for otus}

\item{otu.data}{the data set where each row was a sample and each column was a taxa,}

\item{test.method}{a test including "t" and "wilcoxon"}

\item{cutf}{the significance level}
}
\value{
dif.node the detected differential nodes

p.value the p value for each taxa
}
\description{
An method for phylogeny-aware differential abundance detection
}
\examples{
p <- 29
group <- rep(c(0,1),each=46)
tree <- simulate_tree(p)
ebay.res <- eBay_tree(otu.table,tree=tree,group=group,test.method="t",cutf=0.05)
}
