% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dtm.R
\name{simulation_dtm}
\alias{simulation_dtm}
\title{simulate data from DTM}
\usage{
simulation_dtm(p, seed = 1, N, tree, control_pi, case_pi, control_theta,
  case_theta)
}
\arguments{
\item{p}{the number of taxa}

\item{seed}{a random seed}

\item{N}{sample size of each group}

\item{tree}{a phylogenetic tree}

\item{control_pi}{a probability vector for each taxon which were sum to 1 in control group}

\item{case_pi}{a probability vector for each taxon which were sum to 1 in case group}

\item{control_theta}{the over-dispersion parameter}

\item{case_theta}{the over-dispersion parameter}
}
\value{
dtm_table the otu table
}
\description{
simulate data from DTM
}
